Move utility methods to jobconfig
