+++ robots/pkg/kubevirt/copy/BUILD.bazel
@@ -26,6 +26,7 @@ go_test(
     srcs = ["jobs_test.go"],
     embed = [":go_default_library"],
     deps = [
+        "//robots/pkg/kubevirt/jobconfig:go_default_library",
         "//robots/pkg/querier:go_default_library",
         "@com_github_go_test_deep//:go_default_library",
         "@com_github_google_go_github//github:go_default_library",
